+++ client/pages/goals/_id/index.vue
@@ -100,6 +100,7 @@ export default Vue.extend({
   }
 })
 </script>
+
 <style lang="scss" scoped>
 .content {
   margin: auto;
